Fix bug on number of objects.
When adjacency list of graph has the maximum node index in a right column the following lines result in a error. 

    nents = int(np.array(list(self._weights.keys())).max() + 1)
    assert len(objects) == nents, 'Number of objects do no match'

The reason for that is that `self._weights.keys())).max()` takes a maximum of the first endpoint of an edge. For the datasets, where the maximum label is on the second endpoint of an edge, this results in assertion error. 

For example, if idx has the following structure:
0 1
0 2
The total number of objects equals to 3; while `self._weights.keys())).max() + 1` results in 1. 
Removing assertion resolves the issue.